backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
bm25s==0.2.13
Bottleneck==1.5.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
durationpy==0.9
exceptiongroup==1.3.0
fastapi==0.115.9
ffmpy==0.5.0
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.2
google-auth==2.39.0
googleapis-common-protos==1.70.0
gradio==5.25.2
gradio_client==1.8.0
greenlet==3.2.3
groovy==0.1.2
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==32.0.1
langchain==0.3.23
langchain-chroma==0.2.3
langchain-community==0.3.21
langchain-core==0.3.54
langchain-openai==0.3.14
langchain-pinecone==0.2.8
langchain-tests==0.3.19
langchain-text-splitters==0.3.8
langsmith==0.3.32
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.21.0
openai==1.75.0
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pinecone==7.0.2
pinecone-plugin-assistant==1.6.1
pinecone-plugin-interface==0.0.7
pinecone-text==0.10.0
pluggy==1.6.0
posthog==3.25.0
propcache==0.3.1
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydantic[email]
pydub==0.25.1
Pygments==2.19.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==8.4.0
pytest-asyncio==0.26.0
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
ruff==0.11.6
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
semantic-version==2.10.0
sentence-transformers==2.7.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.45.3
sympy==1.14.0
syrupy==4.9.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.3
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
typer==0.16.0
types-requests==2.32.0.20250602
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
wget==3.2
wrapt==1.17.2
yarl==1.20.0
zipp==3.22.0
zstandard==0.23.0
